refactor seed scripts
